shorten event class name
