+++ src/OneBot/V12/Object/Event/Notice/GroupMemberUnbanEvent.php
@@ -12,7 +12,7 @@ use OneBot\V12\Object\Event\HasOperatorId;
 /**
  * OneBot 群成员解除禁言事件
  */
-class GroupMemberUnbanNoticeEvent extends NoticeEvent
+class GroupMemberUnbanEvent extends NoticeEvent
 {
     use HasGroupId;
     use HasOperatorId;
